pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "working-journal-tracker"
version = "0.1.0"
description = "工作日志追踪工具 - 自动追踪工作时间、生成工作纪要"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sshtunnel>=0.4.0",
    "gitpython>=3.1.0",
    "watchdog>=3.0.0",
    "cryptography>=41.0.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
work-journal = "cli.main:cli"

[tool.setuptools.packages.find]
include = ["cli*", "backend*", "tracker*", "shared*"]
